# api/openapi/swagger.yaml
basePath: /
definitions:
  AdminStoryRequest:
    properties:
      action_label:
        type: string
      action_url:
        type: string
      audience_rules:
        additionalProperties: {}
        type: object
      audience_scope:
        type: string
      caption:
        type: string
      expires_at:
        type: integer
      media_type:
        type: string
      media_url:
        type: string
      priority:
        type: integer
      starts_at:
        type: integer
      thumbnail_url:
        type: string
      title:
        type: string
    type: object
  AdminStoryResponse:
    properties:
      action_label:
        type: string
      action_url:
        type: string
      audience_scope:
        type: string
      caption:
        type: string
      created_at:
        type: integer
      created_by_admin:
        type: string
      expires_at:
        type: integer
      id:
        type: string
      media_type:
        type: string
      media_url:
        type: string
      priority:
        type: integer
      published_at:
        type: integer
      publisher_key:
        type: string
      starts_at:
        type: integer
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: integer
    type: object
...
  City:
    properties:
      city:
        type: string
      createdAt:
        type: integer
      id:
        type: string
      isActive:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      state:
        type: string
      updatedAt:
        type: integer
    type: object
...
  DataResponse:
    properties:
      data: {}
    type: object
...
  ListResponse:
    properties:
      cursor:
        type: string
      data: {}
    type: object
...
  NearbyCityResult:
    properties:
      city:
        type: string
      distanceKm:
        type: number
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      state:
        type: string
    type: object
...
  ProblemDetail:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    email: support@medha.app
    name: Medha Backend Team
    url: https://medha.app
  description: Backend API for the Medha platform. V2 Authentication (OTP-based),
    User Profiles, Events, Matching, and more.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Medha API V2
  version: 2.0.0
paths:
...
  /api/v2/admin/stories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Stories
          schema:
            allOf:
            - $ref: '#/definitions/ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/AdminStoryResponse'
                  type: array
              type: object
      security:
      - AdminToken: []
      summary: List all official stories
      tags:
      - admin-stories
    post:
      consumes:
      - application/json
      parameters:
      - description: Story data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AdminStoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Story created
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminStoryResponse'
              type: object
      security:
      - AdminToken: []
      summary: Create an official story (draft)
      tags:
      - admin-stories
  /api/v2/admin/stories/{id}:
    delete:
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Story deleted
      security:
      - AdminToken: []
      summary: Delete (soft-delete) a story
      tags:
      - admin-stories
    get:
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Story
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminStoryResponse'
              type: object
      security:
      - AdminToken: []
      summary: Get an official story
      tags:
      - admin-stories
    put:
      consumes:
      - application/json
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      - description: Story data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AdminStoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Story updated
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminStoryResponse'
              type: object
      security:
      - AdminToken: []
      summary: Update an official story's content/targeting/window
      tags:
      - admin-stories
  /api/v2/admin/stories/{id}/archive:
    post:
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Story archived
      security:
      - AdminToken: []
      summary: Archive a story
      tags:
      - admin-stories
  /api/v2/admin/stories/{id}/pause:
    post:
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Story paused
      security:
      - AdminToken: []
      summary: Pause a story
      tags:
      - admin-stories
  /api/v2/admin/stories/{id}/publish:
    post:
      parameters:
      - description: Story UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Story published
      security:
      - AdminToken: []
      summary: Publish (activate) a story
      tags:
      - admin-stories
...
  /api/v2/cities/nearby:
    get:
      description: Returns a list of seeded cities within a given radius, sorted by
        distance.
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - description: Radius in km (default 200)
        in: query
        name: radius
        type: integer
      - description: Max results (default 15)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nearby cities
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/NearbyCityResult'
                  type: array
              type: object
        "400":
          description: Missing or invalid parameters
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      summary: Get nearby seeded cities
      tags:
      - cities
  /api/v2/cities/search:
    get:
      description: Searches seeded cities by name. Falls back to MapMyIndia autocomplete
        if not found in database.
      parameters:
      - description: Search query (min 2 characters)
        in: query
        name: q
        required: true
        type: string
      - description: Max results (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching cities
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/City'
                  type: array
              type: object
        "400":
          description: Missing or invalid query parameter
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      summary: Search cities by name
      tags:
      - cities
...
securityDefinitions:
  AdminToken:
    description: Type "Bearer" followed by a space and an admin JWT or break-glass
      token.
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...